Support non-float32 traces in python
Add python support for non-float32 numerics such as int16 and int32.
segyio was always able to read the field that describes the floating
format, but there hasn't been an implementation for anything other than
the 32-bit floating points.

+++ python/segyio/_raw_trace.py
@@ -16,7 +16,7 @@ class RawTrace(object):
             start, stop, step = index.indices(f.tracecount)
             mstart, mstop = min(start, stop), max(start, stop)
             length = max(0, (mstop - mstart + (step - (1 if step > 0 else -1))))
-            buf = np.zeros(shape = (length, len(f.samples)), dtype = np.single)
+            buf = np.zeros(shape = (length, len(f.samples)), dtype = f.dtype)
             l = len(range(start, stop, step))
             buf, _ = self.trace._readtr(start, step, l, buf)
             return buf
